+++ src/commands/helpCommands.ts
@@ -39,7 +39,7 @@ export class HelpCommandsClass extends BaseEditCommandsClass {
         //@@language md
 
         const md_s = `\
-        #Welcome to Leo\'s help system.
+        # Welcome to Leo\'s help system
 
         To learn about **Alt-X** commands, type:
 
@@ -327,13 +327,13 @@ export class HelpCommandsClass extends BaseEditCommandsClass {
         This help discusses only \@file nodes. For other ways of creating
         external files, see:
 
-            [<STARTINGURL>/docs/getting-started/tutorial-scripting](<STARTINGURL>/docs/getting-started/tutorial-scripting) or
-            [<STARTINGURL>/docs/users-guide/directives](<STARTINGURL>/docs/users-guide/directives)
+        [<STARTINGURL>/docs/getting-started/tutorial-scripting](<STARTINGURL>/docs/getting-started/tutorial-scripting) or
+        [<STARTINGURL>/docs/users-guide/directives](<STARTINGURL>/docs/users-guide/directives)
 
         Leo creates external files in an unusual way. Please fee free to ask for
         help:
 
-            https://groups.google.com/forum/#!forum/leo-editor
+        [https://groups.google.com/forum/#!forum/leo-editor](https://groups.google.com/forum/#!forum/leo-editor)
 
         ## Overview